requirements.txt
asgiref==3.5.0
autopep8==1.6.0
coverage==6.3.2
Django==4.0.4
django-filter==21.1
django-shortcuts==1.6
djangorestframework==3.13.1
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
